requirements.txt
mysql-connector-python==8.0.28
psycopg2-binary==2.9.3
